Define 'PiecewisePolynomialParameters' in 'piecewise_polynomial', use 'piecewise_constant' in yaml files instead of ad-hoc implementations in policy functions.
