+++ src/_gettsim/parameters/wohngeld.yaml
@@ -726,42 +726,75 @@ freibetrag_kinder_m:
     reference: Art. 1 G. v. 08.10.2015 BGBl Nr.38, S.1617
     alleinerziehend: 110
     arbeitendes_kind: 100
-freibetrag_behinderung:
+freibetrag_bei_behinderung_gestaffelt_y:
   name:
     de: Jährlicher Freibetrag pro behindertes Haushaltsmitglied
-    en: null
+    en: Annual allowance per disabled household member
   description:
     de: >-
-      Der Betrag ist abhängig vom Behinderungsgrad der jeweiligen Person. Vor 2016 ist
-      der Betrag unterschiedlich für Personen mit Behinderungsgrad über bzw unter 80%.
-      Für Menschen mit Behinderungsgrad unter 100% muss zusätzlich Pflegebedürftigkeit
-      gem. §14 SGB XI. und gleichzeitig Häusliche/teilstationäre/Kurzzeit-Pflege
-      vorliegen. Dies ist aktuell nicht implementiert. § 17 Nr. 1 WoGG. (früher § 17 (2)
-      WoGG)
+      § 17 (2) WoGG
+      Der Betrag ist abhängig vom Behinderungsgrad der jeweiligen Person.
+      Achtung: Die Zahlen vor 1991 erscheinen falsch. Senkung in 1986 unwahrscheinlich,
+      im Zweifel wurde dort eher die Stufe mit Behinderungsgrad 1-80% eingeführt.
     en: null
   unit: Euro
   reference_period: Year
+  type: piecewise_constant
   1984-01-01:
-    0: 0
-    1: 0
-    80: 1224
+    0:
+      lower_threshold: -inf
+      intercept_at_lower_threshold: 0
+    1:
+      lower_threshold: 80
+      upper_threshold: inf
+      intercept_at_lower_threshold: 1224
   1986-01-01:
     deviation_from: previous
-    80: 612
+    1:
+      lower_threshold: 80
+      upper_threshold: inf
+      intercept_at_lower_threshold: 612
   1991-01-01:
-    0: 0
-    1: 1224
-    80: 1536
+    0:
+      lower_threshold: -inf
+      intercept_at_lower_threshold: 0
+    1:
+      lower_threshold: 1
+      intercept_at_lower_threshold: 1224
+    2:
+      lower_threshold: 80
+      upper_threshold: inf
+      intercept_at_lower_threshold: 1536
   2002-01-01:
-    0: 0
-    1: 1200
-    80: 1500
+    0:
+      lower_threshold: -inf
+      intercept_at_lower_threshold: 0
+    1:
+      lower_threshold: 1
+      intercept_at_lower_threshold: 1200
+    2:
+      lower_threshold: 80
+      upper_threshold: inf
+      intercept_at_lower_threshold: 1500
+  # FixMe: Comment in after moving to new infrastructure
+  # 2016-01-01:
+  #   note: Replaced by new rule, see 'freibetrag_bei_behinderung_pauschal_y'
+freibetrag_bei_behinderung_pauschal_y:
+  name:
+    de: Jährlicher Freibetrag pro behindertes Haushaltsmitglied
+    en: Annual allowance per disabled household member
+  description:
+    de: >-
+      § 17 Nr. 1 WoGG.
+      Für Menschen mit Behinderungsgrad unter 100% muss zusätzlich Pflegebedürftigkeit
+      gem. §14 SGB XI. und gleichzeitig Häusliche/teilstationäre/Kurzzeit-Pflege
+      vorliegen. Dies ist aktuell nicht implementiert.
+    en: null
+  unit: Euro
+  reference_period: Year
   2016-01-01:
     reference: 2 Art. 1 G. v. 08.10.2015 BGBl Nr.38 S.1612
     scalar: 1500
-    note: >-
-      neue Regelung - §17 1. WoGG a) GDB = 100 b) GDB < 100 & pflegebedürftig gem. §14
-      SGB XI. + gleichz. Häusliche/teilstationäre/Kurzzeit-Pflege
   2020-01-01:
     reference: Art. 1 G. v. 30.11.2019 BGBl. I S. 1877
     scalar: 1800
