+++ src/ttsim/params-schema.json
@@ -1,102 +1,111 @@
 {
-    "$schema": "http://json-schema.org/draft-07/schema#",
-    "type": "object",
-    "patternProperties": {
-      "^[a-zA-Z0-9äöüßÄÖÜ_]+$": {
-        "type": "object",
-        "properties": {
-          "name": {
-            "type": "object",
-            "properties": {
-              "de": { "type": "string" },
-              "en": { "type": ["string", "null"] }
-            },
-            "required": ["de"],
-            "additionalProperties": false
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "type": "object",
+  "patternProperties": {
+    "^[a-zA-Z0-9äöüßÄÖÜ_]+$": {
+      "type": "object",
+      "properties": {
+        "name": {
+          "type": "object",
+          "properties": {
+            "de": { "type": "string" },
+            "en": { "type": ["string", "null"] }
           },
-          "description": {
-            "type": "object",
-            "properties": {
-              "de": { "type": "string" },
-              "en": { "type": ["string", "null"] }
-            },
-            "required": ["de"],
-            "additionalProperties": false
-          },
-          "unit": {
-            "type": ["string", "null"],
-            "enum": [
-              "Euro",
-              "DM",
-              "Share",
-              "Percent",
-              "Factor",
-              "Year",
-              "Month",
-              "Hour",
-              "Square Meter",
-              "Euro / Square Meter",
-              null
-            ]
-          },
-          "type": {
-            "type": "string",
-            "enum": ["scalar", "dict", "list", "piecewise_linear", "piecewise_quadratic", "require_converter"]
-          },
-          "reference_period": {
-            "type": "string",
-            "enum": ["Year", "Quarter", "Month", "Week", "Day"]
-          },
-          "access_different_date": {
-            "type": "string",
-            "enum": ["vorjahr", "jahresanfang"]
-          }
+          "required": ["de"],
+          "additionalProperties": false
         },
-        "patternProperties": {
-          "^(19|20)[0-9]{2}-(0[1-9]|1[0-2])-[0-3][0-9]$": {
-            "type": "object",
-            "properties": {
-              "value": {
-                "oneOf": [
+        "description": {
+          "type": "object",
+          "properties": {
+            "de": { "type": "string" },
+            "en": { "type": ["string", "null"] }
+          },
+          "required": ["de"],
+          "additionalProperties": false
+        },
+        "unit": {
+          "type": ["string", "null"],
+          "enum": [
+            "Euro",
+            "DM",
+            "Share",
+            "Percent",
+            "Factor",
+            "Year",
+            "Month",
+            "Hour",
+            "Square Meter",
+            "Euro / Square Meter",
+            null
+          ]
+        },
+        "type": {
+          "type": "string",
+          "enum": [
+            "scalar",
+            "dict",
+            "list",
+            "piecewise_constant",
+            "piecewise_linear",
+            "piecewise_quadratic",
+            "piecewise_cubic",
+            "require_converter"
+          ]
+        },
+        "reference_period": {
+          "type": "string",
+          "enum": ["Year", "Quarter", "Month", "Week", "Day"]
+        },
+        "access_different_date": {
+          "type": "string",
+          "enum": ["vorjahr", "jahresanfang"]
+        }
+      },
+      "patternProperties": {
+        "^(19|20)[0-9]{2}-(0[1-9]|1[0-2])-[0-3][0-9]$": {
+          "type": "object",
+          "properties": {
+            "value": {
+              "oneOf": [
                 { "type": ["number"] },
                 { "type": "string", "enum": ["inf", "-inf"] }
               ]
             },
-              "reference": { "type": "string" },
-              "note": { "type": "string" },
-              "deviation_from": { "type": "string" },
-              "min_alter": { "type": "number" },
-              "max_alter": { "type": "number" },
-              "betrag": { "type": "number" },
-              "lower_threshold": {
-                "oneOf": [
-                  { "type": "number" },
-                  { "type": "string", "enum": ["-inf"] }
-                ]
-              },
-              "upper_threshold": {
-                "oneOf": [
-                  { "type": "number" },
-                  { "type": "string", "enum": ["inf"] }
-                ]
-              },
-              "rate": { "type": "number" },
-              "intercept_at_lower_threshold": { "type": "number" }
+            "reference": { "type": "string" },
+            "note": { "type": "string" },
+            "deviation_from": { "type": "string" },
+            "min_alter": { "type": "number" },
+            "max_alter": { "type": "number" },
+            "betrag": { "type": "number" },
+            "lower_threshold": {
+              "oneOf": [
+                { "type": "number" },
+                { "type": "string", "enum": ["-inf"] }
+              ]
             },
-            "additionalProperties": true
+            "upper_threshold": {
+              "oneOf": [
+                { "type": "number" },
+                { "type": "string", "enum": ["inf"] }
+              ]
+            },
+            "rate": { "type": "number" },
+            "intercept_at_lower_threshold": { "type": "number" }
+          },
+          "additionalProperties": true
+        }
+      },
+      "additionalProperties": false,
+      "allOf": [
+        { "required": ["name", "description", "type"] },
+        {
+          "minProperties": 1,
+          "patternProperties": {
+            "^(19|20)[0-9]{2}-(0[1-9]|1[0-2])-[0-3][0-9]$": {}
           }
-        },
-        "additionalProperties": false,
-        "allOf": [
-          { "required": ["name", "description", "type"] },
-          {
-            "minProperties": 1,
-            "patternProperties": {
-              "^(19|20)[0-9]{2}-(0[1-9]|1[0-2])-[0-3][0-9]$": {}
-            }
-          }
-        ]
-      }
-    },
-    "additionalProperties": false
-  }
+        }
+      ]
+    }
+  },
+  "additionalProperties": false
+}
